pyproject.toml
[project]
name = "opensailor.org"
version = "0.1.0"
description = "Open and free sailing database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.3",
    "django>=5.2.1",
    "django-allauth>=65.8.0",
    "django-guardian>=3.0.0",
    "django-htmx>=1.23.0",
    "django-ninja>=1.4.1",
    "django-simple-history>=3.8.0",
    "django-storages[s3]>=1.14.2",
    "boto3>=1.34.69",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "python3-openid>=3.2.0",
    "requests>=2.32.3",
    "python-magic>=0.4.27",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "markdown>=3.8",
]

[tool.setuptools]
package-dir = {"" = "/src/app"}
packages = {find = {where = ["src"], include = ["app*"], exclude = []}}

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "prospector>=1.16.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-it>=0.1.5",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "webapp.test_settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
